TinyMultiHeadLut: fast bf16/fp16/fp32 LUT + alt-carrier bug fix
New TinyMultiHeadLut module (n_anchor_pairs ≤ 15, smooth_mode=False,
n_alternatives=1) with native CUDA kernels and a custom Adam optimizer:

- `tiny_mhlut_backward_na1`: fused weights + carriers backward modeled on
  MHLut's `lprojection_backward_na1_*`, with `AT_DISPATCH_FLOATING_TYPES_AND2`
  so the same kernel handles fp32 / fp16 / bf16. Uses an fp32 atomic-add
  accumulator buffer to avoid bf16/fp16 CAS-loop atomics.
- `embedding_bag` forward (skips the [B, n_lookup_tables, n_outputs]
  intermediate that naive advanced indexing materialises).
- `recompute-in-backward` autograd Function that saves only weights + the
  two int64 lookup index tensors.

Critical bug fix in `_TinyMHLutGatherReduce`: the autograd Function now
threads BOTH `lookup_indices_grad_c` AND `lookup_alt_indices_grad_c` and
returns `grad_main` AND `grad_alt`, matching MHLut's
`lprojection_backward_na1_carriers_kernel`. Without the alt carrier,
`TinyAnchorPairsLookup`'s STE backward computes `du = grad_main *
uncertainty_derivative` instead of `(grad_main - grad_alt) *
uncertainty_derivative`, giving systematically wrong x.grad. This caused a
+0.03 bpb gap to the fp32 baseline that we initially misdiagnosed as bf16
weight ULP / fp16 mantissa precision over 5 fruitless variant runs (v3..v7).

Equivalence test (`test_tiny_vs_full_equivalence.py`) now proves: at the
three exp061 LUT shapes (qk_joint, v_lut, out_proj_largest), Tiny and
MultiHeadLut agree on forward, weights.grad, x.grad to <1e-5, and after 8
SGD steps from identical init the weight drift stays <1e-5.

`TinyMultiHeadLutOptimizer`: AdamW for Tiny weights with configurable
state_dtype / compute_dtype, plus stochastic-rounded fp32→bf16 / fp32→fp16
weight write that defeats the round-to-nearest precision floor on
in-place sub-ULP updates.

`nanochat_exps/exp061_tiny_mhlut/`: fork of exp060 (val=1.7114) measuring
TinyMultiHeadLut at multiple precision configs. v1..v7 archived (all
contaminated by the alt-carrier bug). After the fix, v8 (fp32 Tiny + stock
AdamW) tracks exp060 within VAL noise (|Δbpb|<0.005 at step 2000) and
profiles 1.27× faster (199 ms/step vs 253 ms/step) at 0.68× memory. v9
(bf16 max-hard: bf16 weights + bf16 moments + bf16 compute + SR write)
also tracks exp060 within ±0.006 through step 3800 and profiles 1.24× /
0.63× — small precision tax for an extra 7% memory drop.

+++ src/spiky/lutorch/ranking_tools.py
@@ -514,7 +514,9 @@ class DominanceToVector(nn.Module):
 
     def forward(self, d: torch.Tensor) -> torch.Tensor:
         # d: (..., P). einsum supports any leading dims.
-        v = torch.einsum('...p,kp->...k', d, self.borda_m)
+        # Cast borda_m to d.dtype so bf16/fp16 inputs don't trigger dtype
+        # mismatch (no-op when d is fp32, since `.to(same dtype)` returns self).
+        v = torch.einsum('...p,kp->...k', d, self.borda_m.to(d.dtype))
         if self.ln is not None:
             v = self.ln(v)
         return v
